+++ rulette-mysql-provider/src/main/java/com/github/kislayverma/rulette/mysql/dao/BaseDao.java
@@ -1,0 +1,12 @@
+package com.github.kislayverma.rulette.mysql.dao;
+
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public abstract class BaseDao {
+
+    protected Connection getConnection() throws SQLException, IOException {
+        return DataSource.getInstance(null).getConnection();
+    }
+}
